adding the database management

+++ src/main/java/project/persistence/ShoppingCartRepository.java
@@ -3,6 +3,5 @@ package project.persistence;
 import org.springframework.data.repository.CrudRepository;
 import project.models.ShoppingCart;
 
-public interface ShoppingCartRepository extends CrudRepository<ShoppingCart, Integer> {
-
+public interface ShoppingCartRepository extends CrudRepository<ShoppingCart, Long> {
 }
